+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Client respects Context.Deadline when setting net.Conn deadline.
 - Client listens on Context.Done while waiting for a connection from the pool and returns an error when context context is cancelled.
 - `Query.Column` does not accept relation name any more. Use `Query.Relation` instead which returns an error if relation does not exist.
+- urlvalues package is renamed to urlfilter.
 
 ## v8